RONDB-775: Improved MGET
MGET uses a number of parallel CommittedRead transactions that are sent and
executed in parallel. To avoid having a too large number of transaction objects
we executed those in batches of 100 at a time. This means that each Ndb object
need to be able to handle 100 transactions in parallel.

After finishing a batch we close all transactions, but we retain the result
of the reads in the local buffer we allocated for the query.

Only when all batches are completed do we proceed to put together the result.

Still no handling of large rows implemented.